refactor: jwt package, PasswordEncryptor 이동 [AIR-125]
