+++ aircnc/src/main/java/com/gurudev/aircnc/infrastructure/security/PasswordEncryptor.java
@@ -1,4 +1,4 @@
-package com.gurudev.aircnc.configuration;
+package com.gurudev.aircnc.infrastructure.security;
 
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
