+++ aircnc/src/main/java/com/gurudev/aircnc/infrastructure/security/jwt/Jwt.java
@@ -1,4 +1,4 @@
-package com.gurudev.aircnc.configuration.jwt;
+package com.gurudev.aircnc.infrastructure.security.jwt;
 
 import static lombok.AccessLevel.PRIVATE;
